# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Plomvix API",
    "version": "0.1.0",
    "description": "Plomvix — Indian-built, open-source unified observability and general-purpose database. Supports logs, metrics, telemetry, key-value, and JSON data.",
    "contact": {
      "name": "Plomvix",
      "url": "https://github.com/plomvix/plomvix"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from POST /auth/login"
      },
      "APIKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key from POST /admin/users/{id}/apikey"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "plomvix_token",
        "description": "httpOnly browser session cookie set by POST /auth/login"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "status",
          "error"
        ],
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          },
          "data": {
            "type": "object"
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LoginResponseData": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "GenericSuccessData": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "description": "Authenticate with username and password to receive a JWT token.",
        "operationId": "loginUser",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "username": "admin",
                "password": "changeme"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful. Also sets httpOnly cookie plomvix_token for browser clients",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/LoginResponseData"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout",
        "description": "Invalidates the current JWT token, logging the user out of the session.",
        "operationId": "logoutUser",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "APIKeyAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logout successful. Clears the plomvix_token cookie",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/GenericSuccessData"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh token",
        "description": "Exchange a valid refresh token for a new JWT access token.",
        "operationId": "refreshToken",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "APIKeyAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              },
              "example": {
                "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImFkbWluIiwicm9sZSI6ImFkbWluIn0.dGhpcyBpcyBhIHNhbXBsZSBzaWduYXR1cmU"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully. Refreshes the plomvix_token cookie for browser clients",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/LoginResponseData"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/ingest/logs": {
      "post": {
        "tags": [
          "Ingest"
        ],
        "summary": "Ingest log records",
        "description": "Ingest log records in one of the supported formats: application/json, text/csv, text/x-logfmt, or application/x-syslog.",
        "operationId": "ingestLogs",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "APIKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              },
              "example": {
                "records": [
                  {
                    "level": "info",
                    "message": "Server started",
                    "timestamp": "2026-05-31T12:00:00Z",
                    "fields": {
                      "pid": 1234
                    }
                  }
                ]
              }
            },
            "text/csv": {
              "schema": {
                "type": "string"
              },
              "example": "timestamp,level,message\n2026-05-31T12:00:00Z,info,Server started"
            },
            "text/x-logfmt": {
              "schema": {
                "type": "string"
              },
              "example": "ts=2026-05-31T12:00:00Z level=info msg=\"Server started\" pid=1234"
            },
            "application/x-syslog": {
              "schema": {
                "type": "string"
              },
              "example": "<134>1 2026-05-31T12:00:00Z myhost myapp 1234 ID47 - Server started"
            }
          }
        },
        "responses": {
          "201": {
            "description": "Records ingested successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/IngestResponseData"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or content type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/ingest/json": {
      "post": {
        "tags": [
          "Ingest"
        ],
        "summary": "Ingest JSON records",
        "description": "Ingest arbitrary JSON records. Supports application/json and text/csv content types.",
        "operationId": "ingestJson",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "APIKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              },
              "example": {
                "records": [
                  {
                    "data": {
                      "event": "user_signup",
                      "user_id": "abc123"
                    },
                    "timestamp": "2026-05-31T12:00:00Z"
                  }
                ]
              }
            },
            "text/csv": {
              "schema": {
                "type": "string"
              },
              "example": "timestamp,event,user_id\n2026-05-31T12:00:00Z,user_signup,abc123"
            }
          }
        },
        "responses": {
          "201": {
            "description": "Records ingested successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/IngestResponseData"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or content type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/ingest/kv": {
      "post": {
        "tags": [
          "Ingest"
        ],
        "summary": "Ingest key-value records",
        "description": "Ingest key-value records in JSON format. Supports application/json content type only.",
        "operationId": "ingestKV",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "APIKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              },
              "example": {
                "records": [
                  {
                    "key": "config.app_name",
                    "value": "Plomvix"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Records ingested successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/IngestResponseData"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/query/logs": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Query log records",
        "description": "Query ingested log records with optional filtering. Supports time-range scans.",
        "operationId": "queryLogs",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "description": "Filter expression to narrow results (e.g., level=error)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of records to return",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of records to skip for pagination",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Query results",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/QueryResult"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/query/json": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Query JSON records",
        "description": "Query ingested JSON records with optional filtering. Supports time-range scans.",
        "operationId": "queryJson",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "description": "Filter expression to narrow results",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of records to return",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of records to skip for pagination",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Query results",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/QueryResult"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/query/schema/{type}": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Get inferred schema for a data type",
        "description": "Retrieve the automatically inferred schema for a given data type.",
        "operationId": "getSchemaByType",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "description": "Data type to get schema for",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "logs",
                "metrics",
                "json",
                "kv"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Schema information",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SchemaInfo"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid data type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}